# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/taches": {
      "get": {
        "operationId": "TachesController_lister",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "etat",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "EN_ATTENTE",
                "EN_CORBEILLE",
                "RECLAMEE",
                "APPROUVEE",
                "REJETEE",
                "POST_CLOTURE"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "default": 1
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "taches": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "demandeId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "reference": {
                            "type": "string"
                          },
                          "nomClient": {
                            "type": "string"
                          },
                          "montantTtc": {
                            "type": "number"
                          },
                          "roleCorbeille": {
                            "type": "string"
                          },
                          "ordre": {
                            "type": "number"
                          },
                          "typeActeur": {
                            "type": "string",
                            "enum": [
                              "V",
                              "A",
                              "C"
                            ]
                          },
                          "bloquant": {
                            "type": "boolean"
                          },
                          "slaHeures": {
                            "type": "number"
                          },
                          "modeAffectation": {
                            "type": "string",
                            "enum": [
                              "PULL"
                            ]
                          },
                          "etat": {
                            "type": "string",
                            "enum": [
                              "EN_ATTENTE",
                              "EN_CORBEILLE",
                              "RECLAMEE",
                              "APPROUVEE",
                              "REJETEE",
                              "POST_CLOTURE"
                            ]
                          },
                          "agentClaimId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "dateClaim": {
                            "type": "string",
                            "nullable": true
                          },
                          "verrouExpireAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "echeanceSla": {
                            "type": "string",
                            "nullable": true
                          },
                          "niveauEscalade": {
                            "type": "number"
                          },
                          "dateDecision": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "demandeId",
                          "reference",
                          "nomClient",
                          "montantTtc",
                          "roleCorbeille",
                          "ordre",
                          "typeActeur",
                          "bloquant",
                          "slaHeures",
                          "modeAffectation",
                          "etat",
                          "agentClaimId",
                          "dateClaim",
                          "verrouExpireAt",
                          "echeanceSla",
                          "niveauEscalade",
                          "dateDecision"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "taches",
                    "total"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "taches"
        ]
      }
    },
    "/api/taches/{id}": {
      "get": {
        "operationId": "TachesController_trouver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "demandeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "nomClient": {
                      "type": "string"
                    },
                    "montantTtc": {
                      "type": "number"
                    },
                    "roleCorbeille": {
                      "type": "string"
                    },
                    "ordre": {
                      "type": "number"
                    },
                    "typeActeur": {
                      "type": "string",
                      "enum": [
                        "V",
                        "A",
                        "C"
                      ]
                    },
                    "bloquant": {
                      "type": "boolean"
                    },
                    "slaHeures": {
                      "type": "number"
                    },
                    "modeAffectation": {
                      "type": "string",
                      "enum": [
                        "PULL"
                      ]
                    },
                    "etat": {
                      "type": "string",
                      "enum": [
                        "EN_ATTENTE",
                        "EN_CORBEILLE",
                        "RECLAMEE",
                        "APPROUVEE",
                        "REJETEE",
                        "POST_CLOTURE"
                      ]
                    },
                    "agentClaimId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "dateClaim": {
                      "type": "string",
                      "nullable": true
                    },
                    "verrouExpireAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "echeanceSla": {
                      "type": "string",
                      "nullable": true
                    },
                    "niveauEscalade": {
                      "type": "number"
                    },
                    "dateDecision": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "demandeId",
                    "reference",
                    "nomClient",
                    "montantTtc",
                    "roleCorbeille",
                    "ordre",
                    "typeActeur",
                    "bloquant",
                    "slaHeures",
                    "modeAffectation",
                    "etat",
                    "agentClaimId",
                    "dateClaim",
                    "verrouExpireAt",
                    "echeanceSla",
                    "niveauEscalade",
                    "dateDecision"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "taches"
        ]
      }
    },
    "/api/taches/{id}/claim": {
      "post": {
        "operationId": "TachesController_claim",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "demandeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "nomClient": {
                      "type": "string"
                    },
                    "montantTtc": {
                      "type": "number"
                    },
                    "roleCorbeille": {
                      "type": "string"
                    },
                    "ordre": {
                      "type": "number"
                    },
                    "typeActeur": {
                      "type": "string",
                      "enum": [
                        "V",
                        "A",
                        "C"
                      ]
                    },
                    "bloquant": {
                      "type": "boolean"
                    },
                    "slaHeures": {
                      "type": "number"
                    },
                    "modeAffectation": {
                      "type": "string",
                      "enum": [
                        "PULL"
                      ]
                    },
                    "etat": {
                      "type": "string",
                      "enum": [
                        "EN_ATTENTE",
                        "EN_CORBEILLE",
                        "RECLAMEE",
                        "APPROUVEE",
                        "REJETEE",
                        "POST_CLOTURE"
                      ]
                    },
                    "agentClaimId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "dateClaim": {
                      "type": "string",
                      "nullable": true
                    },
                    "verrouExpireAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "echeanceSla": {
                      "type": "string",
                      "nullable": true
                    },
                    "niveauEscalade": {
                      "type": "number"
                    },
                    "dateDecision": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "demandeId",
                    "reference",
                    "nomClient",
                    "montantTtc",
                    "roleCorbeille",
                    "ordre",
                    "typeActeur",
                    "bloquant",
                    "slaHeures",
                    "modeAffectation",
                    "etat",
                    "agentClaimId",
                    "dateClaim",
                    "verrouExpireAt",
                    "echeanceSla",
                    "niveauEscalade",
                    "dateDecision"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "taches"
        ]
      }
    },
    "/api/taches/{id}/unclaim": {
      "post": {
        "operationId": "TachesController_unclaim",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "demandeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "nomClient": {
                      "type": "string"
                    },
                    "montantTtc": {
                      "type": "number"
                    },
                    "roleCorbeille": {
                      "type": "string"
                    },
                    "ordre": {
                      "type": "number"
                    },
                    "typeActeur": {
                      "type": "string",
                      "enum": [
                        "V",
                        "A",
                        "C"
                      ]
                    },
                    "bloquant": {
                      "type": "boolean"
                    },
                    "slaHeures": {
                      "type": "number"
                    },
                    "modeAffectation": {
                      "type": "string",
                      "enum": [
                        "PULL"
                      ]
                    },
                    "etat": {
                      "type": "string",
                      "enum": [
                        "EN_ATTENTE",
                        "EN_CORBEILLE",
                        "RECLAMEE",
                        "APPROUVEE",
                        "REJETEE",
                        "POST_CLOTURE"
                      ]
                    },
                    "agentClaimId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "dateClaim": {
                      "type": "string",
                      "nullable": true
                    },
                    "verrouExpireAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "echeanceSla": {
                      "type": "string",
                      "nullable": true
                    },
                    "niveauEscalade": {
                      "type": "number"
                    },
                    "dateDecision": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "demandeId",
                    "reference",
                    "nomClient",
                    "montantTtc",
                    "roleCorbeille",
                    "ordre",
                    "typeActeur",
                    "bloquant",
                    "slaHeures",
                    "modeAffectation",
                    "etat",
                    "agentClaimId",
                    "dateClaim",
                    "verrouExpireAt",
                    "echeanceSla",
                    "niveauEscalade",
                    "dateDecision"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "taches"
        ]
      }
    },
    "/api/taches/{id}/approuver": {
      "post": {
        "operationId": "TachesController_approuver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "revue": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "champ": {
                          "type": "string"
                        },
                        "vu": {
                          "type": "boolean"
                        },
                        "correction": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "champ",
                        "vu"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "demandeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "nomClient": {
                      "type": "string"
                    },
                    "montantTtc": {
                      "type": "number"
                    },
                    "roleCorbeille": {
                      "type": "string"
                    },
                    "ordre": {
                      "type": "number"
                    },
                    "typeActeur": {
                      "type": "string",
                      "enum": [
                        "V",
                        "A",
                        "C"
                      ]
                    },
                    "bloquant": {
                      "type": "boolean"
                    },
                    "slaHeures": {
                      "type": "number"
                    },
                    "modeAffectation": {
                      "type": "string",
                      "enum": [
                        "PULL"
                      ]
                    },
                    "etat": {
                      "type": "string",
                      "enum": [
                        "EN_ATTENTE",
                        "EN_CORBEILLE",
                        "RECLAMEE",
                        "APPROUVEE",
                        "REJETEE",
                        "POST_CLOTURE"
                      ]
                    },
                    "agentClaimId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "dateClaim": {
                      "type": "string",
                      "nullable": true
                    },
                    "verrouExpireAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "echeanceSla": {
                      "type": "string",
                      "nullable": true
                    },
                    "niveauEscalade": {
                      "type": "number"
                    },
                    "dateDecision": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "demandeId",
                    "reference",
                    "nomClient",
                    "montantTtc",
                    "roleCorbeille",
                    "ordre",
                    "typeActeur",
                    "bloquant",
                    "slaHeures",
                    "modeAffectation",
                    "etat",
                    "agentClaimId",
                    "dateClaim",
                    "verrouExpireAt",
                    "echeanceSla",
                    "niveauEscalade",
                    "dateDecision"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "taches"
        ]
      }
    },
...
    "/api/taches/{id}/controle": {
      "post": {
        "operationId": "TachesController_soumettreControle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "constat": {
                    "type": "string",
                    "enum": [
                      "CONFORME",
                      "ANOMALIE"
                    ]
                  },
                  "commentaire": {
                    "type": "string"
                  }
                },
                "required": [
                  "constat"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "demandeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "niveau": {
                      "type": "string",
                      "enum": [
                        "FRA",
                        "N1",
                        "N2"
                      ]
                    },
                    "constat": {
                      "type": "string",
                      "enum": [
                        "CONFORME",
                        "ANOMALIE"
                      ]
                    },
                    "commentaire": {
                      "type": "string",
                      "nullable": true
                    },
                    "controleurId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "horodatage": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "demandeId",
                    "niveau",
                    "constat",
                    "commentaire",
                    "controleurId",
                    "horodatage"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "taches"
        ]
      }
    },
...
  },
  "info": {
    "title": "PGD — API",
    "description": "Plateforme de Gestion des Dégrèvements — Orange Côte d'Ivoire",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}